\name{aircraft}

\alias{aircraft}

\title{These data record six characteristics of aircraft designs which appeared during the twentieth century}

\description{
These data record six characteristics of aircraft designs which appeared
during the twentieth century.

The variables are:
\tabular{ll}{
	\code{Yr}     \tab year of first manufacture \cr
	\code{Period} \tab a code to indicate one of three broad time periods \cr
	\code{Power}  \tab total engine power (kW) \cr
	\code{Span}   \tab wing span (m) \cr
	\code{Length} \tab length (m) \cr
	\code{Weight} \tab maximum take-off weight (kg) \cr
	\code{Speed}  \tab maximum speed (km/h) \cr
	\code{Range}  \tab range (km)
}

Source: The data were collected by P. Saviotti and are described in detail
	in Saviotti (1996), "Technological Evolution, Variety and Economy",
	Edward Elgar: Cheltenham.
}

\keyword{smooth}
\keyword{regression}

