\name{airpc}

\alias{airpc}

\title{These data list the first two principal component scores from the
aircraft data, which record six characteristics of aircraft designs 
throughout the twentieth century}

\description{
These data list the first two principal component scores from the
aircraft data, which record six characteristics of aircraft designs 
throughout the twentieth century.

The variables are:
\tabular{ll}{
	\code{Comp.1:} \tab  first principal component score \cr
	\code{Comp.2:} \tab  second principal component score \cr
	\code{Yr:}     \tab  year of first manufacture \cr
	\code{Period:} \tab  a code to indicate one of three broad time periods
}

The data were collected by P. Saviotti and are described in detail
	in Saviotti (1996), "Technological Evolution, Variety and Economy",
	Edward Elgar: Cheltenham.
}

\keyword{smooth}
\keyword{regression}

