\name{birth}

\alias{birth}

\title{Low birthweight in babies}

\description{
The data refer to a study on low birthweight in babies, defined
as "less than 2500 grams", and related factors.

The variables are:
\tabular{ll}{
	\code{Smoke} \tab  indicator of whether the mother is a smoker \cr
	\code{Lwt}   \tab  last menstrual weight of the mother\cr
	\code{Low}   \tab  indicator variable of low weight
}

Hosmer & Lemeshow (1989).  Applied  Logistic Regresion. 
        Wiley, NY.
	The original source contains additional variables; see Appendix 1 of
	Hosmer & Lemeshow for a full list of the data, pp.29-30 and p.92 for 
	additional information.
}

\keyword{smooth}
\keyword{regression}

