\name{bissell}

\alias{bissell}

\title{Flaws in cloth}

\description{
These data refer to the length and the observed number of flaws in
rolls of cloth.

The variables are:
\tabular{ll}{
	\code{Length} \tab length of each roll (m) \cr
	\code{Flaws}  \tab  number of flaws detected
}

Source: Bissell (1972). A negative binomial model with varying element sizes.  Biometrika 59, 435-41.
}

\keyword{smooth}
\keyword{regression}

