\name{bonions}

\alias{bonions}

\title{Yield-density relationship for Brown Imperial Spanish onions}

\description{
These data were colllected in a study of the relationship between the
yield of Brown Imperial Spanish onion plants and the density of planting.

The variables are:
\tabular{ll}{
	\code{Density}  \tab density of planting (plants/m^2) \cr
	\code{Yield}    \tab yield (g/plant) \cr
	\code{Locality} \tab a code to indicate Mount Gambier (1) or Uraidla (2)
}

The data were collected by I.S.Rogers (South Australian Dept. of 
	Agriculture & Fisheries).  They are listed in
	Ratkowsky (1983), Nonlinear Regression Modeling. Dekker, New York.
}

\keyword{smooth}
\keyword{regression}

