\name{britpts}

\alias{britpts}

\title{Coastline of the UK and Ireland}

\description{
These data provide the co-orinates of a set of points lying on the coastlines of the UK and Ireland.

The variables are:
\tabular{ll}{
	\code{britlong} \tab  longitude \cr
	\code{brotlat}  \tab  latitude
}

}

\keyword{smooth}
\keyword{regression}

