\name{citrate}

\alias{citrate}

\title{The relationship between plasma citrate and carbohydrate metabolites}

\description{
These data were collected in an experiment to study the 
relationship between possible daily rhythms of plasma citrate and 
daily rhythms of carbohydrate metabolites during feeding with a 
citrate-poor diet.  During the experiment, plasma citrate 
concentrations were determined for each of 10 subjects at 14 
successive time points during the day.  The measurements covered the 
period 8a.m. to 9p.m. at hourly intervals.  Meals were given at 
8a.m., noon and 5p.m.

The variables are denoted by \code{C08}, ..., \code{C21} and refer to plasma citrate measurements at the indicated hours.

Anderson,A.H., Jensen,E.B. & Schou,G.(1981).  Two-way analysis of variance with correlated errors.  Int.Stat.Rev. 49,153-67.

The data were taken from a report by T.T.Nielsen, N.S.Sorensen and E.B.Jensen.
}

\keyword{smooth}
\keyword{regression}

