\name{coalash}

\alias{coalash}

\title{Coal ash in mining samples}

\description{
These data record the percentage of coal ash found in mining samples 
originally reported by Gomez and Hazen (1970) and subsequently used
by Cressie (1993).

The variables are:
\tabular{ll}{
	\code{East}    \tab  a code for east-west direction \cr
	\code{North}   \tab  a code for north-south direction \cr
	\code{Percent} \tab  the percentage of coalash
}

Cressie, N. (1993). Statistics for Spatial Data, revised edition. New York: Wiley.
    
Gomez, M. and Hazen, K. (1970). Evaluating sulfur and ash distribution in coal seams by statistical response surface regression analysis. Report RI 7377, U.S. Bureau of  Mines, Washington, D. C.
}

\keyword{smooth}
\keyword{regression}

