\name{follicle}

\alias{follicle}

\title{Ovarian follicle counts}

\description{
These data record the number of ovarian follicles, on a log scale, counted from sectioned ovaries of women of various ages.

The variables are:
\tabular{ll}{
	\code{Age}    \tab  age of the woman \cr
	\code{Count}  \tab  follicle count \cr
	\code{Source} \tab  an indicator of the source of the data
}

The data were reported by Block (1952; 1953), Richardson et al. (1987) and A Gougeon.  They are analysed by Faddy & Gosden (1996) and Faddy & Jones (1997).
}

\keyword{smooth}
\keyword{regression}

