\name{geys3d}

\alias{geys3d}

\title{Duration and the time between eruptions for the Old Faithful Geyser}

\description{
These data document the duration of eruptions, and the time between
eruptions, for the Old Faithful Geyser in Yellowstone National Park.

The variables are:
\tabular{ll}{
	\code{Waiting}      \tab  the waiting time before each eruption (minutes) \cr
	\code{Next.waiting} \tab the waiting time following each eruption (minutes) \cr
	\code{Duration}     \tab  the length of an eruption ( minutes)
}

The data were collected by by the Park Geologist, R.A.Hutchinson.  An earlier set of data is reported in Weisberg (1990), Applied Linear Regression, Wiley, New York.  The later set, used here, was reported by Azzalini & Bowman (1990), "A look at some data on the Old Faithful Geyser", Applied Statistics 39, 357-65.
}

\keyword{smooth}
\keyword{regression}

