\name{lcancer}

\alias{lcancer}

\title{Spatial positions of cases of laryngeal cancer}

\description{
These data record the spatial positions of cases of laryngeal cancer
in the North-West of England between 1974 and 1983, together with the
positions of a number of lung cancer patients who were used as controls.
The data have been adjusted to preserve anonymity.

The variables are:
\tabular{ll}{
	\code{Easting}  \tab  a west-east grid reference \cr
	\code{Northing} \tab  a north-south grid reference \cr
	\code{Cancer}   \tab  an indicator of laryngeal (1) or lung (2) cancer
}

Source: Bailey & Gatrell (1995).  Interactive Spatial Data Analysis.  Longman Scientific and Technical, Harlow.  A more extensive set of data is analysed in Kelsall & Diggle, kernel estimation of relative risk, Bernoulli 1, 3-16.
}

\keyword{smooth}
\keyword{regression}

