% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_count.R
\name{LvCnt}
\alias{LvCnt}
\title{Method of Counting Factor Levels in a Dataframe}
\usage{
LvCnt(data, level)
}
\arguments{
\item{data}{a dataframe that you want to explore}

\item{level}{an integer which controls the number of factor level that you want to see. Too many factor levels can convolute the whole understanding.}
}
\value{
A dataframe containing the name of factor levels and counts of each levels
}
\description{
To check several variables in a dataframe, it is important to take grasp the density or table of each variable. This fct is for factor variable, return a dataframe showing factor level and count number of each variable.
}
\examples{
   lvcnt_ex <- LvCnt(iris, 2)
   View(lvcnt_ex)
}
