\name{mackerel}

\alias{mackerel}

\title{The abundance of mackerel eggs}

\description{
These data record the abundance of mackerel eggs off the coast of
north-western Europe, from a multi-country survey in 1992.

The variables are:
\tabular{ll}{
	\code{Density}     \tab  egg density \cr
	\code{mack.lat}    \tab  latitude of sampling position \cr
	\code{mack.long}   \tab  longitude of sampling position \cr
	\code{mack.depth}  \tab  bottom depth \cr
	\code{Temperature} \tab  surface temperature \cr
	\code{Salinity}    \tab  salinity
}

Background to the survey and the data are provided by Watson et al. (1992), Priede and Watson (1993) and Priede et al (1995).  Borchers et al (1997) describe an analysis of the data.
}

\keyword{smooth}
\keyword{regression}

