\name{magrem}

\alias{magrem}

\title{Magnetic remanence}

\description{
These data record measurements of magnetic remanence in specimens of
Precambrian volcanics.

The variables are:
\tabular{ll}{
	\code{maglong} \tab  directional component on a longitude scale \cr
	\code{maglat}  \tab  directional component on a latitude scale
}

Schmidt & Embleton (1985) J.Geophys.Res. 90 (B4), 2967-2984.

The data are also listed in Fisher, Lewis & Embleton (1987), Statistical Analysis of Spherical Data, Cambridge University Press, Cambridge, dataset B6.
}

\keyword{smooth}
\keyword{regression}

