\name{mildew}

\alias{mildew}

\title{Mildew control}

\description{
The data refer to study of mildew control sponsored by Bainbridge,
Jenkyn and Dyke at Rothamsted Experimental Station.  There were four
treatments, one of which was a control.  There were 36 adjacent plots,
with an extra plot at each end.  Nine blocks were created by grouping
the plots in fours.

The variables are:
\tabular{ll}{
	\code{t1, t2, t3}  \tab  indicators of the four treatment groups \cr
	\code{p1, ..., p8} \tab  indicators of the nine blocks\cr
	\code{Yield}       \tab  tons of grain per hectare
}

Draper & Guttman (1980).  Incorporating overlap effects from
	neighbouring units into response surface models.
	Applied Statistics 29, 128-134.
}

\keyword{smooth}
\keyword{regression}

