\name{muscle}

\alias{muscle}

\title{Rat skeletal muscles}

\description{
The data refer to the counts of fibres of different types in 
groups of fibres taken from rat skeletal muscles.  

The variables are:
\tabular{ll}{
	\code{row.labels} \tab  indicators of the four treatment groups \cr
	\code{TypeI.R}    \tab  indicators of the nine blocks \cr
	\code{TypeI.P}    \tab  tons of grain per hectare \cr
	\code{TypeI.B}    \tab  tons of grain per hectare \cr
	\code{TypeII}     \tab  tons of grain per hectare \cr
}

Hand, Daly, Lunn, McConway and Ostrowski (1994).  A handbook of Small Data Sets.  Chapman & Hall: London.  The data were collected by M. Khan and M. Khan.
}

\keyword{smooth}
\keyword{regression}

