\name{nile}

\alias{nile}

\title{Water level of the River Nile}

\description{
These data record historical data on the water level of the River Nile.

The variables are:
\tabular{ll}{
	\code{Volume} \tab  Annual volume of the Nile River (discharge at Aswan, 10^8 m^3) \cr
	\code{Year}   \tab  1871-1970
}

Cobb, G. (1978). The problem of the Nile: conditional solution to a change-point problem.  Biometrika 65, 243-251.
}

\keyword{smooth}
\keyword{regression}

