\name{pause}
\alias{pause}
\title{
Pause before continuing execution  
}
\description{
If a program produces several plots, isertion of \code{pause()} between
two plots suspends execution until the <Enter> key is pressed, to
allow inspection of the current plot.
}
\usage{
pause()
}
% Converted by Sd2Rd version 0.3-2.
