\name{poles}

\alias{poles}

\title{Positions of the south pole}

\description{
These data refer to positions of the south pole determined from the
palaeomagnetic study of New Caledonian laterites.

The variables are:
\tabular{ll}{
	\code{Latitude} \cr
	\code{Longitude}
}

The data were collected by Falvey and Musgrave.	They are listed in Fisher, Lewis & Embleton (1987), Statistical Analysis of Spherical Data, Cambridge University Press, Cambridge, dataset B1.
}

\keyword{smooth}
\keyword{regression}

