\name{radioc}

\alias{radioc}

\title{Radiocarbon in Irish oak}

\description{
These data record high precision measurements of radiocarbon on Irish oak, used to construct a calibration curve.

The variables are:
\tabular{ll}{
	\code{Rc.age}    \tab  age predicted from the radiocarbon dating process \cr
	\code{Precision} \tab  a measure of precision of the dating process \cr
	\code{Cal.age}   \tab  true calendar age
}

Pearson & Qua (1993).  High precision 14Cmeasurement of Irish oaks to show the natural 14C variations from AD 1840 - 5000 BC: a correction.  Radiocarbon 35, 105-123.
}

\keyword{smooth}
\keyword{regression}

