% Auto-generated: do not edit by hand
\name{sideMenu}

\alias{sideMenu}

\title{SideMenu component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
sideMenu(id=NULL, items=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{items}{Unnamed list. An array of id, label}

\item{className}{Character. Navbar class name}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

