\name{smacker}

\alias{smacker}

\title{Mackerel data from a Spanish survey}

\description{
These data were recorded by a Spanish survey, as part of a multi-country 
survey of the abundance of mackerel eggs off the coast of north-western 
Europe, in 1992.  

The variables are:
\tabular{ll}{
	\code{Density}     \tab  egg density \cr
	\code{smack.lat}   \tab  latitude of sampling position \cr
	\code{smack.long}  \tab  longitude of sampling position \cr
	\code{smack.depth} \tab  bottom depth \cr
	\code{Temperature} \tab  surface temperature \cr
	\code{D200}        \tab  distance from the 200m depth contour line
}

Background to the survey and the data are provided by Watson et al. (1992), Priede and Watson (1993) and Priede et al (1995).  Borchers et al (1997) describe an analysis of the data.
}

\keyword{smooth}
\keyword{regression}

