\name{stanford}

\alias{stanford}

\title{Survival times from the Stanford Heart Transplant Study}

\description{
These data refer to the survival times of patients of different ages
from the Stanford Heart Transplant Study.

The variables are:
\tabular{ll}{
	\code{Log.time} \tab  egg density \cr
	\code{Status}   \tab  latitude of sampling position \cr
	\code{Age}      \tab  longitude of sampling position
}

Source: Miller & Halpern (1980).  Regression with censored data.  Biometrika 69, 521-531.
}

\keyword{smooth}
\keyword{regression}

