\name{tephra}

\alias{tephra}

\title{Tephra layer}

\description{
These data record the percentages of aluminium oxide found in samples
from a tephra layer resulting from a volcanic eruption in Iceland
around 3500 years ago.

The variables are:
\tabular{ll}{
	\code{Al2O3}  \tab  percentage of aluminium oxide
}

The data were collected by A.Dugmore.  The geological background to the data is given by Dugmore et al (1992), Geochemical stability of finegrained silicic tephra in Iceland and Scotland, J.Quatern.Sci. 7, 173-83.
}

\keyword{smooth}
\keyword{regression}

