\name{trawl}

\alias{trawl}

\title{Trawl data from the Great Barrier Reef}

\description{
These data refer to a survey of the fauna on the sea bed lying
between the coast of northern Queensland and the Great Barrier Reef.
The sampling region covered a zone which was closed to commercial
fishing, as well as neighbouring zones where fishing was permitted.

The variables are:
\tabular{ll}{
	\code{Zone}      \tab  an indicator for the closed (1) and open (0) zones \cr
	\code{Year}      \tab  an indicator of 1992 (0) or 1993 (1) \cr
	\code{Latitude}  \tab  latitude of the sampling position \cr
	\code{Longitude} \tab  longitude of the sampling position \cr
	\code{Depth}     \tab  bottom depth \cr
	\code{Score1}    \tab  catch score 1 \cr
	\code{Score2}    \tab  catch score 2
}

The details of the survey and an analysis of the data are provided by Poiner et al. (1997), The effects of prawn trawling in the far northern section of the Great Barrier Reef, CSIRO Division of Marine Research, Queensland Dept. of Primary Industries.
}

\keyword{smooth}
\keyword{regression}

