\name{worm}

\alias{worm}

\title{Human parasitic worm infections}

\description{
These data record the occurrence of a human parasitic worm infection in residents of a rural community in China. 

The variables are:
\tabular{ll}{
	\code{Age}       \tab  age of the resident \cr
	\code{Infection} \tab  presence (1) or absence (0) of infection \cr
	\code{Sex}       \tab  male (1) or female (2)
}

The background to the data, and an analysis, are described by Weidong et al. (1996), Ascaris, people and pigs in a rural community of Jiangxi province, China, Parasitology 113, 545-57.
}

\keyword{smooth}
\keyword{regression}

