\name{pause}
\alias{pause}
\title{
Pause before continuing execution  
}
\description{
If a program produces several plots on the same window, insertion of
\code{pause()} between two plots suspends execution until the <Enter>
key is pressed, to allow inspection of the current plot.
}
\usage{
pause()
}
\keyword{misc}
% Converted by Sd2Rd version 1.15.
