% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM.R
\name{SM}
\alias{SM}
\title{Make a Histogram, Boxplot and Statistical Analysis}
\usage{
SM()
}
\arguments{
\item{x}{numeric variable}
}
\description{
Make a Histogram, Boxplot and Statistical Analysis
}
